Updating doxyfiles for new version, updating multi_funct and mm_funct tests, fixed mm_funct_strings, work on long double integrators.

+++ src/base/multi_funct.h
@@ -97,7 +97,7 @@ namespace o2scl {
     template<class vec_t=boost::numeric::ublas::vector<double> >
       double operator()(size_t nv, const vec_t &x) {
 
-      for(int i=0;i<nv;i++) {
+      for(size_t i=0;i<nv;i++) {
 	vars[st_vars[i]]=x[i];
       }
